recomendasi mantap

+++ app/src/main/java/com/munaz/nutrisiapp/data/request/RekomendasiReq.kt
@@ -1,13 +1,16 @@
 package com.munaz.nutrisiapp.data.request
 
+import okhttp3.RequestBody
+import retrofit2.http.Part
+
 data class RekomendasiReq(
-    val usia_user: Float,
-    val berat_badan: Float,
-    val tinggi_badan: Float,
+    val usia_user: Int,
+    val berat_badan: Int,
+    val tinggi_badan: Int,
     val jenis_kelamin: String,
     val riwayat_penyakit: String,
     val riwayat_alergi: String,
     val faktor_aktivitas: String,
     val faktor_stress: String,
-    val dislike_food: String?
+    val dislike_food: String ?
 )
